+++ Modules/Intent.Modules.ModuleBuilder.TypeScript/release-notes.md
@@ -1,3 +1,7 @@
+### Version 3.3.9
+
+- Update: Partial Class Access modifier will now be `public` when `Custom` is used as Templating method.
+
 ### Version 3.3.8
 
 - Updated dependencies and supported client versions to prevent warnings when used with Intent Architect 4.x.